requirements.txt

itsdangerous
Jinja2
MarkupSafe
Werkzeug>=0.14
wheel
gunicorn
Flask-OAuthlib
flask_mail
